+++ CHANGES.md
@@ -50,7 +50,7 @@
 
   - <https://github.com/georust/gdal/pull/439>
 
-- Added Histogram calculation
+- Added histogram calculation
 
   - <https://github.com/georust/gdal/pull/468>